backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
bcrypt
python-multipart
pydantic
pydantic-settings
google-genai
python-dotenv
aiosmtplib
email-validator
cryptography
httpx
PyPDF2
pytesseract
Pillow
